extract middleware supporting message bus from command bus and event bus implementations
